requirements.txt
streamlit==1.23.1
pandas==2.0.1
numpy==1.23.5
pillow==.5.0
torchvision==0.15.2
torch==2.0.1
opencv-python==4.7.0.72